WIP refactor field rendering code out of preset editor
