+++ modules/ui/index.js
@@ -13,6 +13,7 @@ export { uiEditMenu } from './edit_menu';
 export { uiEntityEditor } from './entity_editor';
 export { uiFeatureInfo } from './feature_info';
 export { uiFeatureList } from './feature_list';
+export { uiField } from './field';
 export { uiFlash } from './flash';
 export { uiFullScreen } from './full_screen';
 export { uiGeolocate } from './geolocate';
